fix: seed test expected int not int64

+++ option_test.go
@@ -121,7 +121,7 @@ func TestParseSeed(t *testing.T) {
 	a := assert.New(t)
 
 	seed, err := ParseSeed(100)
-	a.Equal(100, seed, "seed should be 100")
+	a.Equal(int64(100), seed, "seed should be 100")
 	a.NoError(err, "there should be no error")
 }
